Refactor events and archive utils
